Separate files for CKey and CExtKey classes
