+++ _bitcoin_common.vcxproj.filters
@@ -19,6 +19,9 @@
     <Filter Include="Source Files\key">
       <UniqueIdentifier>{db6f8a17-7d14-49e0-beb2-493547e7f656}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Source Files\key\secret">
+      <UniqueIdentifier>{dd096edd-b052-4997-8a37-c8c0ed821ef5}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\chainparams.cpp">
@@ -306,14 +309,8 @@
     <ClCompile Include="src\base58.cpp">
       <Filter>Source Files\key</Filter>
     </ClCompile>
-    <ClCompile Include="src\key.cpp">
-      <Filter>Source Files\key</Filter>
-    </ClCompile>
     <ClCompile Include="stdafx.cpp" />
     <ClCompile Include="src\BitcoinAddress.cpp">
-      <Filter>Source Files\key</Filter>
-    </ClCompile>
-    <ClCompile Include="src\BitcoinSecret.cpp">
       <Filter>Source Files\key</Filter>
     </ClCompile>
     <ClCompile Include="src\BitcoinExtKeyBase.cpp">
@@ -330,6 +327,15 @@
     </ClCompile>
     <ClCompile Include="src\pubkey.cpp">
       <Filter>Source Files\key</Filter>
+    </ClCompile>
+    <ClCompile Include="src\key.cpp">
+      <Filter>Source Files\key\secret</Filter>
+    </ClCompile>
+    <ClCompile Include="src\BitcoinSecret.cpp">
+      <Filter>Source Files\key\secret</Filter>
+    </ClCompile>
+    <ClCompile Include="src\ExtKey.cpp">
+      <Filter>Source Files\key\secret</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -681,14 +687,8 @@
     <ClInclude Include="src\base58.h">
       <Filter>Source Files\key</Filter>
     </ClInclude>
-    <ClInclude Include="src\key.h">
-      <Filter>Source Files\key</Filter>
-    </ClInclude>
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="src\BitcoinAddress.h">
-      <Filter>Source Files\key</Filter>
-    </ClInclude>
-    <ClInclude Include="src\BitcoinSecret.h">
       <Filter>Source Files\key</Filter>
     </ClInclude>
     <ClInclude Include="src\BitcoinExtKeyBase.h">
@@ -705,6 +705,15 @@
     </ClInclude>
     <ClInclude Include="src\pubkey.h">
       <Filter>Source Files\key</Filter>
+    </ClInclude>
+    <ClInclude Include="src\key.h">
+      <Filter>Source Files\key\secret</Filter>
+    </ClInclude>
+    <ClInclude Include="src\BitcoinSecret.h">
+      <Filter>Source Files\key\secret</Filter>
+    </ClInclude>
+    <ClInclude Include="src\ExtKey.h">
+      <Filter>Source Files\key\secret</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
